[fix] plugin system fails on missing properties
when setting 'plugin.enabled' to true in 'addon-self-administration.properties'
but not providing values for 'plugin.jar.path' or 'plugin.classname' an
exception was thrown at runtime. now such misconfiguration will be
detected at deployment time, failing the deployment process early. due
to this changes the plugin will be loaded after bean initialization
instead of when the first call is made to the plugin handler.