Remove explicit dependency on Jackson
After we restored compatibility of the connector with older Jackson
version, we can now remove the explicit dependency on Jackson. Spring
Boot already defines a working Jackson version that we should use. This
way we reduce the complexity of the POM and also get automatic updates
with new Spring Boot releases instead of relying on the explicit
version, which will be outdated in the future.