Publish latest snapshot to Bintray
This commit changes that only one snapshot artifact is available
on Bintray and remove the deployment to the oss jfrog snapshot
maven repo. This artifact does not need to be available as maven
dependency and also not the future releases, so this commit removes
the maven release plugin and the maven profile 'integration'.