Stop packaging of test configuration file and assets
The usage of maven-resources-plugin in the profile for the ITs, led to
accidental packaging of the addon-administration.properties and some
test assets into the WAR file. This in turn broke the complete app, as
the packaged config file has a higher precedence than the one in
/etc/osiam. Also, the packages test assets made it impossible to
overwrite them in `/etc/osiam` too.

Solve the issue by copying the config file and test assets into the
exploded WAR file in the Docker container.